split module in sql interface
